requirements.txt
email-validator
libsql-client
python-decouple
python-multipart
python-jose[cryptography]